pyproject.toml
[tool.poetry]
name = "Renoir"
version = "1.3.1"
description = "A templating engine designed with simplicity in mind"
authors = ["Giovanni Barillari <gi0baro@d4net.org>"]
license = "BSD-3-Clause"

readme = "README.md"
homepage = "https://github.com/emmett-framework/renoir"
repository = "https://github.com/emmett-framework/renoir"

keywords = ["templates", "html", "web"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML"
]

packages = [
    {include = "renoir"},
    {include = "tests", format = "sdist"}
]
include = [
    "CHANGES.md",
    "LICENSE",
    "docs/**/*"
]

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/emmett-framework/renoir/issues"

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pylint = "^2.4.4"
pyyaml = "^5.3.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
